LPD-32234 Moved Util to correct location
