+++ modules/apps/commerce/commerce-order-content-web/src/main/java/com/liferay/commerce/order/content/web/internal/fragment/renderer/OrderItemsDataSetFragmentRenderer.java
@@ -7,7 +7,7 @@ package com.liferay.commerce.order.content.web.internal.fragment.renderer;
 
 import com.liferay.commerce.model.CommerceOrder;
 import com.liferay.commerce.order.content.web.internal.constants.CommerceOrderFragmentFDSNames;
-import com.liferay.commerce.order.content.web.internal.util.CommerceOrderInfoItemUtil;
+import com.liferay.commerce.order.content.web.internal.item.util.CommerceOrderInfoItemUtil;
 import com.liferay.commerce.product.url.CPFriendlyURL;
 import com.liferay.commerce.service.CommerceOrderService;
 import com.liferay.fragment.model.FragmentEntryLink;
